+++ services/product-service/src/main/java/com/abdelrahman/productservice/repository/ProductInventoryRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ProductInventoryRepository extends JpaRepository<ProductInventory,Integer> {
+public interface ProductInventoryRepository extends JpaRepository<ProductInventory,Long> {
 }
